refactor: 폴더 구조 리팩토링 - layout, card types 분리
