--- apps/client/app/types/card.ts
@@ -1,14 +1,0 @@
-export interface CardItem {
-  id: number;
-  imageUrl: string;
-  title: string;
-  category: string;
-  author: string;
-}
-
-export interface CardProps extends Omit<CardItem, "id"> {}
-
-export interface CardGridProps {
-  items: CardItem[];
-  limit?: number;
-}
