+++ apps/client/components/common/Card/Card.types.ts
@@ -1,0 +1,14 @@
+export interface CardItem {
+	id: number;
+	imageUrl: string;
+	title: string;
+	category: string;
+	author: string;
+}
+
+export interface CardProps extends Omit<CardItem, "id"> {}
+
+export interface CardGridProps {
+	items: CardItem[];
+	limit?: number;
+}
